Move local_run command to `cli`
