+++ tests/cli/test_local_run.py
@@ -6,16 +6,18 @@ from pathlib import Path
 
 import pytest
 
-from jobrunner import local_run
+from jobrunner.cli import local_run
 from jobrunner.lib.subprocess_utils import subprocess_run
+
+
+FIXTURE_DIR = Path(__file__).parents[1].resolve() / "fixtures"
 
 
 @pytest.mark.slow_test
 @pytest.mark.needs_docker
 def test_local_run(tmp_path):
-    project_fixture = str(Path(__file__).parent.resolve() / "fixtures/full_project")
     project_dir = tmp_path / "project"
-    shutil.copytree(project_fixture, project_dir)
+    shutil.copytree(str(FIXTURE_DIR / "full_project"), project_dir)
     local_run.main(project_dir=project_dir, actions=["analyse_data"])
     assert (project_dir / "output/input.csv").exists()
     assert (project_dir / "counts.txt").exists()
@@ -30,9 +32,8 @@ def test_local_run(tmp_path):
     not os.environ.get("STATA_LICENSE"), reason="No STATA_LICENSE env var"
 )
 def test_local_run_stata(tmp_path, monkeypatch):
-    project_fixture = str(Path(__file__).parent.resolve() / "fixtures/stata_project")
     project_dir = tmp_path / "project"
-    shutil.copytree(project_fixture, project_dir)
+    shutil.copytree(str(FIXTURE_DIR / "stata_project"), project_dir)
     monkeypatch.setattr("jobrunner.config.STATA_LICENSE", os.environ["STATA_LICENSE"])
     local_run.main(project_dir=project_dir, actions=["stata"])
     env_file = project_dir / "output/env.txt"
