+++ tests/lib/test_log_utils.py
@@ -2,7 +2,7 @@ import logging
 import time
 from datetime import datetime
 
-from jobrunner import local_run
+from jobrunner.cli import local_run
 from jobrunner.lib import log_utils
 from jobrunner.models import Job, JobRequest
